fix: 코드 리뷰 반영(사용하지 않는 구문 삭제 및 아이콘 lucide 사용)
