+++ src/app/(general)/signup/page.tsx
@@ -6,14 +6,14 @@ import Link from 'next/link'
 export default function SignupPage() {
   return (
     <>
-      <div className="flex flex-1 items-center justify-center px-4 py-12">
+      <div className="flex min-w-lg items-center justify-center">
         <div className="w-full max-w-md">
-          <div className="mb-8 text-center">
+          <div className="mb-8 flex w-full flex-col items-center justify-center">
             <h1 className="mb-2 text-2xl font-bold text-gray-900">회원가입</h1>
             <p className="text-sm text-gray-500">Welearn과 함께 학습을 시작하세요.</p>
           </div>
 
-          <div className="rounded-xl border border-gray-200 bg-white p-8 shadow-sm">
+          <div className="w-full rounded-xl border border-gray-200 bg-white p-8 shadow-sm">
             <SignupForm />
           </div>
